Update and rename JiugeYang_BCO.json to Jiuge_Yang_BCO000530.json
I added the author contribution type with the addition of "derived from" and put the doi of the publication in the usability domain. I also tried to add the review status in the usability domain however the way I did it looks different from the specification docs. Please let me know if my current BCO looks fine and if not, what to revise. Thank you very much!