Update Elizabeth_Lee_BCO_000548
Hello, my BCO was already approved on Wednesday 10/16 by @jpat1546.  The comments from @jpat1546 mentioned that 2 changes need to be made to the Provenance Domain prior to submitting to pFDA.  I made the 2 changes to the Provenance Domain in this version of my BCO.  In addition, I added xref to the Description Domain, and I added additional information to the Execution Domain.  I will submit this version to the pFDA.  Thank you!